+++ src/app/modules/+pos/view/default/index.ts
@@ -7,7 +7,7 @@ import {PosDefaultSalesCheckoutBottomBarComponent} from "./sales/checkout/bottom
 import {PosDefaultSalesCheckoutListComponent} from "./sales/checkout/list.component";
 import {PosDefaultSalesCheckoutGridProductImageComponent} from "./sales/checkout/grid/product-image.component";
 import {PosDefaultSalesCheckoutGrandTotalComponent} from "./sales/checkout/grand-total.component";
-import {PosDefaultSalesOutletRegisterComponent} from "./sales/outlet-register.component";
+import {PosDefaultSalesOutletRegisterComponent} from "./outlet-register.component";
 import {PosDefaultSalesCheckoutCartComponent} from "./sales/checkout/cart.component";
 import {PosDefaultSalesCheckoutCartCustomersComponent} from "./sales/checkout/cart/customers.component";
 import {PosDefaultSalesCheckoutActionBarComponent} from "./sales/checkout/actions-bar.component";
