+++ src/app/modules/+pos/view/default/outlet-register.component.html
